Zero stack buffers before using them.
This is still broken. The ZeroMemory() macro tries to get the length
of buffers by sizeof()ing them, but this is wrong since it's really
sizeof()ing a pointer.

One of the other commits actually fixing this issue [1] should be
merged upstream. This is only an ugly workaround in the mean time.

[1]: https://github.com/lanurmi/QR-Image-embedded/commit/c1c9864